Fix: Makes font variants work with font fallbacks
Changes the font cache from an Arc<Font> to an Arc<RwLock<Font>>
so that we can get mutable references to fonts during font fallback
iteration to be able to set the font variant.

Also adds a new SimpleFontFallbackIter which we use instead
of the more complicated cosmic_text. This lets us try to find
a glyph in every loaded font. This works well for us since we only
ever load a collection of font subsets.